+++ Source/ImoutoRebirth.Arachne/ImoutoRebirth.Arachne.Infrastructure/Abstract/IBooruAvailabilityChecker.cs
@@ -1,0 +1,6 @@
+﻿namespace ImoutoRebirth.Arachne.Infrastructure.Abstract;
+
+public interface IBooruAvailabilityChecker
+{
+    Task<bool> IsAvailable(CancellationToken ct);
+}