+++ Source/ImoutoRebirth.Arachne/ImoutoRebirth.Arachne.Infrastructure/LoaderFabrics/GelbooruLoaderFabric.cs
@@ -3,6 +3,7 @@ using Imouto.BooruParser;
 using Imouto.BooruParser.Implementations.Gelbooru;
 using ImoutoRebirth.Arachne.Core.Models;
 using ImoutoRebirth.Arachne.Infrastructure.Abstract;
+using ImoutoRebirth.Arachne.Infrastructure.Availability;
 using Microsoft.Extensions.Options;
 
 namespace ImoutoRebirth.Arachne.Infrastructure.LoaderFabrics;
@@ -18,4 +19,7 @@ internal class GelbooruLoaderFabric : IBooruLoaderFabric
     public IBooruApiLoader Create() => new GelbooruApiLoader(
         _flurlClientFactory,
         Options.Create(new GelbooruSettings() { PauseBetweenRequestsInMs = 1 }));
+
+    public IBooruAvailabilityChecker CreateAvailabilityChecker()
+        => new SimpleAvailabilityChecker(_flurlClientFactory, new Uri("https://gelbooru.com"));
 }