+++ Source/ImoutoRebirth.Arachne/ImoutoRebirth.Arachne.Infrastructure/LoaderFabrics/Rule34LoaderFabric.cs
@@ -3,6 +3,7 @@ using Imouto.BooruParser;
 using Imouto.BooruParser.Implementations.Rule34;
 using ImoutoRebirth.Arachne.Core.Models;
 using ImoutoRebirth.Arachne.Infrastructure.Abstract;
+using ImoutoRebirth.Arachne.Infrastructure.Availability;
 using Microsoft.Extensions.Options;
 
 namespace ImoutoRebirth.Arachne.Infrastructure.LoaderFabrics;
@@ -18,4 +19,7 @@ internal class Rule34LoaderFabric : IBooruLoaderFabric
     public IBooruApiLoader Create() => new Rule34ApiLoader(
         _flurlClientFactory,
         Options.Create(new Rule34Settings { PauseBetweenRequestsInMs = 1 }));
+
+    public IBooruAvailabilityChecker CreateAvailabilityChecker()
+        => new SimpleAvailabilityChecker(_flurlClientFactory, new Uri("https://rule34.xxx"));
 }